error middleware

+++ src/controllers/artworkControllers.js
@@ -1,55 +1,54 @@
 const database = require('../config/db');
+const { BadRequestError, NotFoundError } = require('../utils/errors');
+const { isValidArtwork, isValidRequest } = require('../utils/validate-request');
 
 // =========== GET ROUTES ===========
-const getArtworks = async (request, response) => {
-    // TODO: validate page and limit
+const getArtworks = async (request, response, next) => {
     let { page, limit } = request.query;
-    limit = parseInt(limit);
-    offset = limit * parseInt(page);
-
-    if (!limit || !page) return response.status(400).send('Define pagination');
-
     try {
+        if (!isValidRequest(page, limit)) throw new BadRequestError('Pagination missing');
+
+        limit = parseInt(limit);
+        const offset = limit * parseInt(page);
+
         const { rows, rowCount } = await database().query('SELECT * FROM artworks LIMIT $1 OFFSET $2', [limit, offset]);
-        if (rowCount === 0) return response.sendStatus(404);
+        if (rowCount === 0) throw new NotFoundError;
+        
         response.status(200).json(rows);
     } catch (error) {
-        console.log(error);
-        return response.sendStatus(500);
+        next(error)        
     };    
 };
 
-const getArtworkById = async (request, response) => {
+const getArtworkById = async (request, response, next) => {
     const { id } = request.params;
     try {
         const { rows, rowCount } = await database().query('SELECT * FROM artworks WHERE id=$1', [id]);
-        if (rowCount === 0) return response.sendStatus(404);
+        if (rowCount === 0) throw new NotFoundError;
         response.status(200).json(rows);
     } catch (error) {
-        console.log(error);
-        return response.sendStatus(500);
-    }
+        next(error)        
+    };
 };
 
-const getPeriodsOfArtwork = async (request, response) => {
+const getPeriodsOfArtwork = async (request, response, next) => {
     const { id } = request.params;
     try {
         const { rows, rowCount } = await database().query('SELECT name FROM periods WHERE id IN (SELECT period_id FROM periods_artworks WHERE artwork_id=$1)', [id]);
-        if (rowCount === 0) return response.sendStatus(404);
+        if (rowCount === 0) throw new NotFoundError;
         response.status(200).json(rows);
     } catch (error) {
-        console.log(error);
-        return response.sendStatus(500);
+        next(error)       
     };
 };
 
 // =========== PUT ROUTES ===========
-const updateArtwork = async (request, response) => {
+const updateArtwork = async (request, response, next) => {
     const { id } = request.params;
     const { file, artist, year, title } = request.body;
     try {
         const { rows, rowCount } = await database().query('SELECT * FROM artworks WHERE id=$1', [id]);
-        if (rowCount === 0) return response.sendStatus(404);
+        if (rowCount === 0) throw new NotFoundError;
 
         const artwork = rows[0];
 
@@ -64,58 +63,56 @@ const updateArtwork = async (request, response) => {
         response.status(200).send(`Artwork modified with ID: ${id}`);
 
     } catch (error) {
-        console.log(error);
-        return response.sendStatus(500);
+        next(error)       
     };
 };
 
-const putPeriodInArtwork = async (request, response) => {
+const putPeriodInArtwork = async (request, response, next) => {
     const { artwork_id, period_id } = request.params;
     try {
+        if (!isValidRequest(artwork_id, period_id)) throw new BadRequestError('Fields missing');
         await database().query('INSERT INTO periods_artworks VALUES($1, $2)', [artwork_id, period_id]);
         return response.sendStatus(201);
     } catch (error) {
-        console.log(error);
-        if (error.code === '23505') return response.status(400).send('Resource already exists');
-        else return response.sendStatus(500);
+        next(error)
     };
 };
 
 // =========== POST ROUTES ===========
-// The postArtwork works with multipart form data. It happens because we need to upload the corresponding image file for the artwork when we submit it.
+// The postArtwork works with multipart form data, because we need to upload the corresponding image file for the artwork when we submit it.
 
-const postArtwork = async (request, response) => {
-    // TODO: handle multer errors first
-    const filename = request.file.originalname;
+const postArtwork = async (request, response, next) => {
     const { title, artist, year } = request.body;
     try {
+        const filename = request.file.originalname;
+        if (!isValidArtwork(request.body)) throw new BadRequestError('Fields missing');
         await database().query('INSERT INTO artworks (file, artist, year, title) VALUES ($1, $2, $3, $4)', [filename, artist, year, title]);
         return response.sendStatus(201);
     } catch (error) {
-        console.log(error);
-        return response.sendStatus(500);
+        next(error)
     };
 };
 
-const deleteArtwork = async (request, response) => {
+// =========== DELETE ROUTES ===========
+const deleteArtwork = async (request, response, next) => {
     const { id } = request.params;
     try {
         await database().query('DELETE FROM artworks WHERE id=$1', [id]);
         return response.sendStatus(200);
     } catch (error) {
-        console.log(error);
-        return response.sendStatus(500);
+        next(error)
     };
 };
 
-const deletePeriodFromArtwork = async (request, response) => {
+const deletePeriodFromArtwork = async (request, response, next) => {
     const { artwork_id, period_id } = request.params;
     try {
+        if (!isValidRequest(artwork_id, period_id)) throw new BadRequestError('Fields missing');
+
         await database().query('DELETE FROM periods_artworks WHERE artwork_id=$1 AND period_id=$2', [artwork_id, period_id]);
         return response.sendStatus(200);
     } catch (error) {
-        console.log(error);
-        return response.sendStatus(500);
+        next(error)
     };
 };
